+++ Runtime/components/DestroyInProductionBuild.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Utilities.Components
+namespace IKhom.UtilitiesLibrary.Runtime.components
 {
     public class DestroyInProductionBuild : MonoBehaviour
     {