+++ Runtime/components/SafeArea.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Utilities.Components
+namespace IKhom.UtilitiesLibrary.Runtime.components
 {
     [DisallowMultipleComponent]
     public class SafeArea : MonoBehaviour