+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2024-07-27
+### Added
+- Reorganize classes and small changes to assemblies
+
+## [1.0.1] - 2024-07-08
+### Added
+- Added utility script to get specified classes from assemblies
+
 ## [1.0.0] - 2024-06-28
 ### Added
 - First release of most used extensions methods